+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## [Unreleased]
 ### Added
-- Added Account import (moved from `alpha-ui`)
+- Added view `quicksight.summary` with the latest state of the current day
+### Changed
+### Fixed
+### Removed
+### -
+
+
+## [1.0.2] - 2025-06-20 11:51 UTC
+### Added
+- Added view `quicksight.summary` with the latest state of the current day
 ### Changed
 - Updated dependencies
 ### Fixed